DeleteComponentCommand minor cleanup/rename. (#43289)
commit
